requirements.txt
cv2
discord.py
numpy
python-dotenv
requests
thefuzz